Fix inclusion of React, Lodash & others
