+++ projects/plugins/jetpack/extensions/blocks/map/utils/get-map-provider.js
@@ -1,0 +1,10 @@
+import { isAtomicSite, isSimpleSite } from '../../../../../../js-packages/shared-extension-utils';
+
+const getMapProvider = () => {
+	if ( isAtomicSite() || isSimpleSite() || window.location.search.includes( 'mapkit' ) ) {
+		return 'mapkit';
+	}
+	return 'mapbox';
+};
+
+export default getMapProvider;
